monkeybook/backend/requirements.txt
# Needed for UserMixin
Flask-Login
